Add definition for generic object type `t_generic_obj`
This is to allow object types to have function pointers to
"member functions" and to allow calling such "member functions"
with the same form of statement.

All object types will need to have 4 function pointers pointing to
their respective translate, rotate, resize_1 and resize_2
functions as the first 4 members within the struct.
The pointers should be assigned during object creation, or at
least before the functions are called.

As an example, if `obj_ptr` is a pointer to an object such as
`t_plane`, the `rotate` function of `t_plane` can be called with
a statement similar to the one below:
`((t_generic_obj *)obj_ptr)->rotate(obj_ptr, vect(10,0,0));`

The same statement would also work for `t_cylinder` and other
compatible object types.