Save object pointer for each pixel on image in `mlxs` struct
This allows retrieving of object pointers directly based on the
pixel being clicked on.
The pointer can then be used to manipulate the object.

An alternative to this would be to find the nearest intersect
between the ray and the objects again, which would take up more
compute time, escpecially when there are many objects in the
scene.