Free pointer returned by `mlx_init` during cleanup
I believe that it should be the responsibility of minilibx to free
the pointer when `mlx_destroy_display` is called as the memory
allocation was done by `mlx_init`, but the pointer was never
freed. Even though I don't think this is a good idea, I am freeing
the pointer just so that there won't be a memory leak being
reported.